backend/requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==5.2.9
djangorestframework==3.16.1
pillow==12.1.0
reportlab==4.4.9
sqlparse==0.5.5
